analytics/requirements.txt
requests
pandas
google-analytics-data
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit
